idt gate refactoring

+++ src/kernel/arch/x86_64/include/panic.h
@@ -10,5 +10,5 @@ void PrintRegs(isr_state_t *reg);
 NORET UNSAFE void DebugPageFault(void* addr);
 
 #define SystemRaiseHardError(reason, desc) _SystemRaiseHardError(__FILE__, __LINE__, reason, desc, NULL);
-#define ISR_SystemRaiseHardError(desc, regs) _SystemRaiseHardError(NULL, (int)NULL, "ISR_STOP", desc, regs);
+#define ISR_SystemRaiseHardError(desc, regs) _SystemRaiseHardError(NULL, (uint64_t)NULL, "ISR_STOP", desc, regs);
 #define DebugRaiseKernelPanic() _SystemRaiseHardError(__FILE__, __LINE__, "DEBUG_STOP", "Manually invoked panic", NULL)
